Moving rest_api and CodeExplorer under apps folder in delphi directory
