+++ delphi/apps/CodeExplorer/.gitignore
