+++ delphi/apps/CodeExplorer/run.sh
