+++ delphi/apps/CodeExplorer/setup_sista_vm.sh
