+++ delphi/apps/CodeExplorer/static/codemirror/lib/codemirror.js
