+++ delphi/apps/CodeExplorer/static/codemirror/theme/monokai.css
