+++ delphi/apps/CodeExplorer/static/hyde.css
