+++ delphi/apps/CodeExplorer/templates/cyjs.js
