+++ delphi/apps/CodeExplorer/templates/editor.js
