+++ delphi/apps/CodeExplorer/templates/navbar.html
