+++ delphi/apps/rest_api/icm_api_skeleton_generator.py
