+++ delphi/apps/rest_api/setup_sista_vm.sh
