+++ delphi/apps/rest_api/utils.py
@@ -1,12 +1,12 @@
 from uuid import uuid4
-from delphi.icm_api.models import (
+from delphi.apps.rest_api.models import (
     Evidence,
     ICMMetadata,
     CausalVariable,
     CausalRelationship,
     DelphiModel,
 )
-from delphi.icm_api import create_app, db
+from delphi.apps.rest_api import create_app, db
 from datetime import date
 from delphi.random_variables import LatentVar
 import numpy as np
